+++ src/common/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod generate_nearest_n;
 pub(crate) mod generate_nearest_one;
+pub(crate) mod generate_within;
