+++ src/fixed/query/nearest_n.rs
@@ -4,7 +4,7 @@ use std::ops::Rem;
 
 use crate::fixed::kdtree::{Axis, KdTree};
 use crate::fixed::neighbour::Neighbour;
-use crate::types::{Content, Index, is_stem_index};
+use crate::types::{is_stem_index, Content, Index};
 
 use crate::generate_nearest_n;
